fix: moves ports into sensor and actuators only
