+++ src/functionalities/Sensor.js
@@ -1,16 +1,9 @@
-const Functionality = require('./Functionality')
+const PhysicalFunctionality = require('./PhysicalFunctionality')
 
-class Sensor extends Functionality {
+class Sensor extends PhysicalFunctionality {
 
   static get TYPE() {
     return 'Sensor'
-  }
-
-  constructor(functionalityData) {
-    super({
-      isVirtual: false,
-      ...functionalityData,
-    })
   }
 
 }
